Add files via upload
This is an alternative design I developed for the Town Hall window of the game Banished, including some new features. The design is unique but the style and icons are taken directly from the game Banished for symmetry and familiarity convenience, icons copyright to Shining Rock Software. The modding tool kit doesn't work on my computer so I can't implement it myself but I'm hoping to find some people interested in working on the mod with me. This was made in Numbers, which is Apple's version of Excel.